split up and rename set_up_distributed_site
now all functions that are called in core_logic on the central site are
also called in core_logic on the remote site.

this way set_up_distributed_site can be renamed to
connect_central_to_remote because this is what it does.

sadly it's not easily possible to move all the calls in a common
function (configure_site) and call this for the central and the remote,
as then we would have to call activate changes twice on the central
site.

Change-Id: I23ac0c795db0580194b19b69a8e7653351025851